requirements.txt
scrapy==2.14.1
scrapy-selenium==0.0.7
selenium==4.41.0
fastapi==0.131.0
uvicorn==0.41.0
pymongo==4.16.0
python-dotenv==1.2.1
